moved pouchdb functions to pouchdb file
